Move modal to separate component
